分支从 main 当时的尖端 `aae2794` 切出，合并期间 main 没动：合并结果的树与
`sr/15-cut-over` 的树逐字节相同，闸门在那棵树上跑满。

+++ src/session/shell/details.rs
@@ -85,7 +85,7 @@ pub(super) fn draw(canvas: &mut Canvas<'_>, session: &Session, area: Rect, narro
         .position(|row| row.choice == Some(session.views.config.choice))
         .unwrap_or(0);
     let viewport = Viewport::with_margin(rows.len(), usize::from(shown), at);
-    let from = usize::from(viewport.from());
+    let from = viewport.from();
     for (i, row) in rows.iter().enumerate().skip(from).take(usize::from(shown)) {
         canvas.line(
             area.x + 2,
